be ready to go back - and review in detail

+++ arch/cpu/cpu.c
@@ -28,7 +28,7 @@ void cpu_print_psr(uint32_t psr) {
     kprintf("%c%c%c%c %c %c%c%c", n, z, c, v, e, i, f, t);
     
     kprintf(" %s", mode);
-    kprintf(" 0x%08x", psr);   
+    kprintf(" 0x%08x", (unsigned int) psr);   
 }
 
 bool is_user_mode(const uint32_t cpsr) {
