+++ kernel/exception_print.c
@@ -60,7 +60,7 @@ static void print_mode_regs(const char *name, uint32_t mode,
         spsr = cpu_get_banked_spsr(mode);
     }
     
-    kprintf("\n%s| LR: 0x%08x | SP: 0x%08x | SPSR: ", name, lr, sp);
+    kprintf("\n%s| LR: 0x%08x | SP: 0x%08x | SPSR: ", name, (unsigned int) lr, (unsigned int) sp);
     cpu_print_psr(spsr);
 }
 void print_exception_infos(enum exc_kind kind, const struct exc_frame* frame) {
@@ -72,35 +72,35 @@ void print_exception_infos(enum exc_kind kind, const struct exc_frame* frame) {
         (kind == EXC_IRQ)  ? "IRQ" : "INVALID";
 
     kprintf("############ EXCEPTION ############\n");
-    kprintf("%s an Adresse: 0x%08x\n", name, frame->lr);
+    kprintf("%s an Adresse: 0x%08x\n", name, (unsigned int) frame->lr);
     
     if (kind == EXC_DABT) {
         uint32_t dfsr = mmu_get_dfsr();
         uint32_t dfar = mmu_get_dfar();
         const char* desc = get_fsr_description(dfsr);
-        kprintf("Data Fault Status Register: 0x%08x -> %s\n", dfsr, desc);
-        kprintf("Data Fault Adress Register: 0x%08x\n", dfar);
+        kprintf("Data Fault Status Register: 0x%08x -> %s\n", (unsigned int) dfsr, desc);
+        kprintf("Data Fault Adress Register: 0x%08x\n", (unsigned int) dfar);
     }
     
     if (kind == EXC_PABT) {
         uint32_t ifsr = mmu_get_ifsr();
         uint32_t ifar = mmu_get_ifar();
         const char* desc = get_fsr_description(ifsr);
-        kprintf("Instruction Fault Status Register: 0x%08x -> %s\n", ifsr, desc);
-        kprintf("Instruction Fault Adress Register: 0x%08x\n", ifar);
+        kprintf("Instruction Fault Status Register: 0x%08x -> %s\n", (unsigned int) ifsr, desc);
+        kprintf("Instruction Fault Adress Register: 0x%08x\n", (unsigned int) ifar);
     }
     
     kprintf("\n>> Registerschnappschuss <<\n");
-    kprintf("R0: 0x%08x  R5: 0x%08x  R10: 0x%08x\n", frame->r[0], frame->r[5], frame->r[10]);
-    kprintf("R1: 0x%08x  R6: 0x%08x  R11: 0x%08x\n", frame->r[1], frame->r[6], frame->r[11]);
-    kprintf("R2: 0x%08x  R7: 0x%08x  R12: 0x%08x\n", frame->r[2], frame->r[7], frame->r[12]);
-    kprintf("R3: 0x%08x  R8: 0x%08x\n", frame->r[3], frame->r[8]);
-    kprintf("R4: 0x%08x  R9: 0x%08x\n", frame->r[4], frame->r[9]);
+    kprintf("R0: 0x%08x  R5: 0x%08x  R10: 0x%08x\n", (unsigned int) frame->r[0], (unsigned int) frame->r[5], (unsigned int) frame->r[10]);
+    kprintf("R1: 0x%08x  R6: 0x%08x  R11: 0x%08x\n", (unsigned int) frame->r[1], (unsigned int) frame->r[6], (unsigned int) frame->r[11]);
+    kprintf("R2: 0x%08x  R7: 0x%08x  R12: 0x%08x\n", (unsigned int) frame->r[2], (unsigned int) frame->r[7], (unsigned int) frame->r[12]);
+    kprintf("R3: 0x%08x  R8: 0x%08x\n", (unsigned int) frame->r[3], (unsigned int) frame->r[8]);
+    kprintf("R4: 0x%08x  R9: 0x%08x\n", (unsigned int) frame->r[4], (unsigned int) frame->r[9]);
     
     kprintf("\n>> Modusspezifische Register <<\n");
     
     kprintf("User/System | LR: 0x%08x | SP: 0x%08x | CPSR: ",
-            cpu_get_banked_lr(CPU_USR), cpu_get_banked_sp(CPU_USR));
+            (unsigned int) cpu_get_banked_lr(CPU_USR), (unsigned int) cpu_get_banked_sp(CPU_USR));
     cpu_print_psr(frame->spsr);
     
     print_mode_regs("IRQ         ", CPU_IRQ, kind == EXC_IRQ, frame);
